Put bittrex and binance api calls and db handling into their own files
